validation-service/requirements.txt
# Validation Service Requirements
# Generated from pyproject.toml dependencies

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Code Quality Tools
ruff>=0.1.6
radon>=6.0.1
bandit>=1.7.6

# Test Framework Support
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Coverage Tools
coverage>=7.3.0

# CLI Framework
typer>=0.9.0
rich>=13.7.0

# Utilities
httpx>=0.25.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
jsonschema>=4.20.0

# Security Scanning
safety>=2.3.5
